Simplifying a bit the sidebarBoardItem code
